pyproject.toml
[tool.poetry]
name = "nomad-airb"
version = "0.1.0"
description = ""
authors = ["MIT"]
readme = "README.md"
packages = [{include = "nomad_airb"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.3"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
